requirements.txt
python-dotenv
grequests
aiofiles